+++ vm/src/numeric/modulo.rs
@@ -2,7 +2,7 @@
 //!
 //! The Modulo trait can be used for getting the modulo (instead of remainder)
 //! of a number.
-use rug::Integer;
+use num_bigint::BigInt;
 
 pub trait Modulo<RHS = Self> {
     type Output;
@@ -44,24 +44,24 @@ impl Modulo for i128 {
     }
 }
 
-impl<'a> Modulo<&'a Integer> for Integer {
-    type Output = Integer;
+impl<'a> Modulo<&'a BigInt> for BigInt {
+    type Output = BigInt;
 
     fn modulo(self, rhs: &Self) -> Self::Output {
         ((self % rhs) + rhs) % rhs
     }
 }
 
-impl Modulo<i32> for Integer {
-    type Output = Integer;
+impl Modulo<i32> for BigInt {
+    type Output = BigInt;
 
     fn modulo(self, rhs: i32) -> Self::Output {
         ((self % rhs) + rhs) % rhs
     }
 }
 
-impl Modulo for Integer {
-    type Output = Integer;
+impl Modulo for BigInt {
+    type Output = BigInt;
 
     fn modulo(self, rhs: Self) -> Self::Output {
         self.modulo(&rhs)
